requirements.txt
python-telegram-bot==21.4
supabase
google-generativeai
groq